Q2/requirements.txt
django
python-dotenv
psycopg2
djangorestframework
drf-yasg